requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0
django-bootstrap-datepicker-plus==5.0.5
django-unfold==0.18.1
dotty-dict==1.3.1
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-gitlab==4.2.0
python-semantic-release==8.7.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
shellingham==1.5.4
smmap==5.0.1
sqlparse==0.4.4
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
gunicorn